+++ versioninfo.h
@@ -1,5 +1,5 @@
 //generated resource header 
 #define GEN_LATEST_VERSION_STRING "1.0.0\0" 
-#define GEN_COMMIT_STRING "041b8f4a\0" 
-#define GEN_COMMIT_COUNT_STRING "533\0" 
-#define GEN_VERSION_STRING "1.0.0.533\0" 
+#define GEN_COMMIT_STRING "f9a7babd\0" 
+#define GEN_COMMIT_COUNT_STRING "541\0" 
+#define GEN_VERSION_STRING "1.0.0.541\0"